backend/requirements.txt
celery

Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-jwt

gunicorn

psycopg2-binary

redis

asgiref==3.3.4
Markdown==3.3.4
Pygments==2.8.1
pytz==2021.1
sqlparse==0.4.1